# Conflicts:
#	Debug/CS230 - .99C44468.tlog/CL.command.1.tlog
#	Debug/CS230 - .99C44468.tlog/CL.read.1.tlog
#	Debug/CS230 - .99C44468.tlog/CL.write.1.tlog
#	Debug/CS230 - .99C44468.tlog/link.command.1.tlog
#	Debug/CS230 - .99C44468.tlog/link.read.1.tlog
#	Debug/CS230 - .99C44468.tlog/link.write.1.tlog
#	Debug/CS230 - Simple Demo.exe
#	Debug/vc140.idb

+++ EnemyAIMeleeBehavior.c
@@ -101,7 +101,7 @@ void EnemyAI_Melee_AttackUpdate(GameObject* enemy, EnemyContainer* enemyContaine
 
   if (enemyContainer->attackWindup > 0)
   {
-    enemyContainer->attackWindup -= AEFrameRateControllerGetFrameTime();
+    enemyContainer->attackWindup -= (float)AEFrameRateControllerGetFrameTime();
   }
 
   if (enemyContainer->attackWindup <= 0)
@@ -138,7 +138,7 @@ void EnemyAI_Melee_CooldownUpdate(GameObject* enemy, EnemyContainer* enemyContai
   enemy->physics->velocity.y = 0;
   if (enemyContainer->attackCooldown > 0)
   {
-    enemyContainer->attackCooldown -= AEFrameRateControllerGetFrameTime();
+    enemyContainer->attackCooldown -= (float)AEFrameRateControllerGetFrameTime();
   }
 
   float distanceToPlayer = Vector2DSquareDistance(&(enemy->physics->position), &(enemy->target->physics->position));
